wip - calling Rust from hypothesis-ruby
