+++ hypothesis-ruby/Gemfile
@@ -10,3 +10,4 @@ gem 'rubocop', '~> 0.51.0'
 gem 'simplecov', '~> 0.15.1'
 gem 'yard', '~> 0.9.12'
 gem 'redcarpet', '~> 3.4.0'
+gem 'rutie', '~> 0.0.3'
