+++ hypothesis-ruby/hypothesis-specs.gemspec
@@ -17,6 +17,6 @@ DESCRIPTION
   s.homepage    = 'https://github.com/HypothesisWorks/hypothesis/tree/master/hypothesis-ruby'
   s.license     = 'MPL-2.0'
   s.extensions = Dir['ext/extconf.rb']
-  s.add_dependency 'helix_runtime', '~> 0.7.0'
+  s.add_dependency 'rutie', '~> 0.0.3'
   s.add_runtime_dependency 'rake', '>= 10.0', '< 13.0'
 end
